feat(food): Add FatSecret API for food search (application-services)
Integrates the FatSecret API to fetch food data. This includes services
for authentication with token caching and for searching products by text
query, barcode, and ID.

- Implements `FoodApiService` and `FatSecretTokenService`.
- Adds `FoodApiSettings` for API configuration.
- Introduces new DTOs (`FoodProductDto`, `FoodSearchDto`) for API responses.
- Updates domain models and DTOs to use `double` for nutrient values.
- Replaces `Quantity` with `Amount` and `ServingUnit` for more detail.
- Secures `FoodApiController` endpoints with authorization.

BREAKING CHANGE: The `FoodItem` value object and DTOs have changed.
Nutrient properties are now `double`. `Quantity` is replaced by `Amount`
and `ServingUnit`. API endpoints return new DTOs.

+++ src/ForgeFit.Domain/Aggregates/FoodAggregate/FoodEntry.cs
@@ -27,10 +27,10 @@ public class FoodEntry : Entity, ITimeFields
     }
 
     public Guid UserId { get; private set; }
-    public int Calories { get; private set; }
-    public int Carbs { get; private set; }
-    public int Protein { get; private set; }
-    public int Fat { get; private set; }
+    public double Calories { get; private set; }
+    public double Carbs { get; private set; }
+    public double Protein { get; private set; }
+    public double Fat { get; private set; }
     public DayTime DayTime { get; private set; }
     public DateTime Date { get; private set; }
     public DateTime CreatedAt { get; init; }
